managed to get the security config to activate via profile removing the property hack from the code and property files small config refactoring
