--- src/main/java/com/management/store/rest/config/EnableSecurityConfig.java
@@ -1,11 +1,0 @@
-package com.management.store.rest.config;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Profile;
-import org.springframework.security.config.annotation.method.configuration.EnableGlobalMethodSecurity;
-
-@Configuration
-@Profile("secure")
-@EnableGlobalMethodSecurity(prePostEnabled = true)
-public class EnableSecurityConfig {
-}
