+++ src/main/java/com/management/store/rest/config/MethodSecurityConfiguration.java
@@ -1,0 +1,50 @@
+package com.management.store.rest.config;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Profile;
+import org.springframework.security.config.annotation.method.configuration.EnableGlobalMethodSecurity;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.core.userdetails.User;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.provisioning.InMemoryUserDetailsManager;
+import org.springframework.security.web.SecurityFilterChain;
+
+@Configuration
+@Profile("secure")
+@EnableGlobalMethodSecurity(prePostEnabled = true)
+public class MethodSecurityConfiguration {
+
+    @Bean
+    public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
+        http
+                .authorizeHttpRequests((authz) -> authz
+                        .antMatchers("/error", "/info")
+                        .permitAll()
+                        .anyRequest()
+                        .authenticated()
+                )
+                .csrf().disable()
+                .httpBasic();
+        return http.build();
+    }
+
+    @Bean
+    public UserDetailsService users() {
+        // The builder will ensure the passwords are encoded before saving in memory
+        User.UserBuilder users = User.withDefaultPasswordEncoder();
+        UserDetails user = users
+                .username("basic")
+                .password("basic")
+                .roles("BASIC")
+                .build();
+        UserDetails admin = users
+                .username("admin")
+                .password("admin")
+                .roles("BASIC", "ADMIN")
+                .build();
+        return new InMemoryUserDetailsManager(user, admin);
+
+    }
+}
