+++ src/main/resources/application-secure.properties
@@ -1,7 +1,4 @@
 server.port=8090
-#
-#Endpoint security config
-run.with.security=true
 
 #JPA properties
 spring.jpa.generate-ddl=true
