+++ src/main/resources/application.properties
@@ -1,7 +1,4 @@
 server.port=8090
-#
-#Endpoint security config
-run.with.security=false
 
 #JPA properties
 spring.jpa.generate-ddl=true
