+++ src/test/resources/application-secure.properties
@@ -4,6 +4,3 @@ spring.datasource.password=test
 
 spring.jpa.generate-ddl=true
 spring.jpa.hibernate.ddl-auto=create-drop
-
-#disable security
-run.with.security=true
